Extract format command to separate file
