+++ src/cli/src/commands/index.ts
@@ -1,3 +1,4 @@
+export * from "./format.js";
 export * from "./generate-feather-metadata.js";
 export * from "./generate-gml-identifiers.js";
 export * from "./generate-quality-report.js";
